python-service/requirements-portfolio.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
asyncpg>=0.29.0
aiohttp>=3.9.0
httpx>=0.27.0
pandas>=2.0.3
beautifulsoup4>=4.12.0
pydantic>=2.5.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
apscheduler>=3.10.4
anthropic>=0.40.0
sendgrid>=6.11.0
